added tree generation, no collision tho. intergrated mine and matei's setting menu and cleaned some code. added coord visualisation to the swing menu.

+++ core/src/com/mygdx/game/main/DataField.java
@@ -17,7 +17,7 @@ public class DataField
     public static double[] targetRXY;
     public static double[] coordinatesandVelocity;
     public static double[] sandPit = {50,50,50,50};
-    public static float x;
-    public static float y;
+    public static double x;
+    public static double y;
     public static Forest gameForest;
 }
